+++ java/wotb-web/src/main/java/com/wotb/web/boost/repository/BoostRequestAssignmentRepository.java
@@ -17,5 +17,7 @@ public interface BoostRequestAssignmentRepository extends JpaRepository<BoostReq
 
     List<BoostRequestAssignment> findByBoosterIdAndUnassignedAtIsNull(Long boosterId);
 
+    List<BoostRequestAssignment> findByBoosterIdOrderByAssignedAtDesc(Long boosterId);
+
     long countByBoosterIdAndUnassignedAtIsNull(Long boosterId);
 }